# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseInputRequest": {
        "additionalProperties": false,
        "description": "Operator input captured immutably at case creation.",
        "properties": {
          "content": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Content",
            "type": "object"
          },
          "input_kind": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Input Kind",
            "type": "string"
          }
        },
        "required": [
          "input_kind",
          "content"
        ],
        "title": "CaseInputRequest",
        "type": "object"
      },
      "CaseListResponse": {
        "additionalProperties": false,
        "description": "Stable tenant-visible case list.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CaseListResponse",
        "type": "object"
      },
      "CaseResponse": {
        "additionalProperties": false,
        "description": "Tenant-authorized current case projection.",
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_actor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Actor Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "state",
          "version",
          "title",
          "category",
          "severity",
          "asset_id",
          "created_by_actor_id",
          "created_at",
          "updated_at"
        ],
        "title": "CaseResponse",
        "type": "object"
      },
...
      "CaseState": {
        "description": "States of the case projection, in the order they can normally progress.",
        "enum": [
          "new",
          "investigating",
          "diagnosed",
          "fix_proposed",
          "needs_information",
          "confirmed",
          "resolved",
          "learned"
        ],
        "title": "CaseState",
        "type": "string"
      },
...
      "ConfigPreviewRequest": {
        "additionalProperties": false,
        "description": "Raw config stays request-local and is never recorded in an event.",
        "properties": {
          "config": {
            "description": "Cisco-style configuration paste for redaction.",
            "minLength": 1,
            "title": "Config",
            "type": "string"
          }
        },
        "required": [
          "config"
        ],
        "title": "ConfigPreviewRequest",
        "type": "object"
      },
      "ConfigPreviewResponse": {
        "additionalProperties": false,
        "description": "Redacted-only derivative; raw config is neither persisted nor disclosed.",
        "properties": {
          "redacted_content": {
            "title": "Redacted Content",
            "type": "string"
          },
          "redacted_content_sha256": {
            "title": "Redacted Content Sha256",
            "type": "string"
          },
          "redaction_version": {
            "title": "Redaction Version",
            "type": "string"
          },
          "report": {
            "$ref": "#/components/schemas/RedactionReportResponse"
          }
        },
        "required": [
          "redacted_content",
          "redacted_content_sha256",
          "redaction_version",
          "report"
        ],
        "title": "ConfigPreviewResponse",
        "type": "object"
      },
      "CreateCaseRequest": {
        "additionalProperties": false,
        "description": "Client-supplied case intent; identity/scope remain server-derived.",
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseInputRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "severity"
        ],
        "title": "CreateCaseRequest",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Stable error envelope returned by every API failure.",
        "properties": {
          "code": {
            "examples": [
              "validation_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "examples": [
              "The submitted request is invalid."
            ],
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "examples": [
              "0d8049a8-ea40-4407-b1f4-690b28ee5d46"
            ],
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Top-level error response schema.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "JsonValue": {},
      "RedactionReportResponse": {
        "additionalProperties": false,
        "properties": {
          "redacted_line_count": {
            "title": "Redacted Line Count",
            "type": "integer"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RedactionRuleSummaryResponse"
            },
            "title": "Rules",
            "type": "array"
          },
          "source_line_count": {
            "title": "Source Line Count",
            "type": "integer"
          }
        },
        "required": [
          "source_line_count",
          "redacted_line_count",
          "rules"
        ],
        "title": "RedactionReportResponse",
        "type": "object"
      },
      "RedactionRuleSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "line_count": {
            "title": "Line Count",
            "type": "integer"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "line_count",
          "occurrence_count"
        ],
        "title": "RedactionRuleSummaryResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OIDC access token": {
        "description": "A signed OIDC access token issued for the NetOps API audience.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NetOps Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases": {
      "get": {
        "description": "List a filtered, cursor-paginated case queue visible to the signed scope.",
        "operationId": "list_cases_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "low",
                    "medium",
                    "high",
                    "critical"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A signed access token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The signed user lacks case permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "OIDC access token": []
          }
        ],
        "summary": "List Cases",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "description": "Create a case once, recording immutable creation lineage and an outbox event.",
        "operationId": "create_case_v1_cases_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "Idempotent replay of an existing case."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A signed access token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The signed user lacks case permission."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "OIDC access token": []
          }
        ],
        "summary": "Create Case",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/cases/{case_id}/config-preview": {
      "post": {
        "description": "Return a bounded redacted Cisco preview without persistence or event emission.",
        "operationId": "preview_case_config_v1_cases__case_id__config_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A signed access token is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The signed user lacks case permission."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Config paste exceeds preview limits."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "OIDC access token": []
          }
        ],
        "summary": "Preview Case Config",
        "tags": [
          "config previews"
        ]
      }
    },
...
  }
}